Put this logic specifically in the page transformer
